vibration/silent-ignore.html

<!DOCTYPE html>
<html xmlns='http://www.w3.org/1999/xhtml' lang='en'>
  <head>
    <meta charset='utf-8'/>
    <title>Vibration API: test that calls to vibrate() are silently ignored when the device cannot vibrate</title>
    <link rel='author' title='Robin Berjon' href='mailto:robin@berjon.com'/>
    <link rel='help' href='http://www.w3.org/TR/vibration/#methods'/>
    <meta name='flags' content='vendor-prefix, dom, no-vibrator'/>
    <meta name='assert' content='If the device does not provide a vibration mechanism, or it is disabled, the user agent must silently ignore any invocations of the vibrate() method.'/>
    <!-- 
      When you review this test, please add
      <link rel='reviewer' title='Your Name' href='mailto:your-email-address'/>
    -->
    <link rel='stylesheet' href='/resources/testharness.css' media='all'/>
  </head>
  <body>
    <h1>Description</h1>
    <p>
      <strong>This test is only useful on devices that do not have vibration capability</strong>.
      If your device supports vibration, then <strong>skip</strong> this test. An implementation
      supporting this API but running on a device that cannot vibrate must silently ignore the
      call (we test that it doesn't throw).
    </p>
    <div id='log'></div>
    <script src='/resources/testharness.js'></script>
    <script src='support/feature-detection.js'></script>
    <script src='support/helpers.js'></script>
    <script>
      if (supportsVibration()) {
	  test(function () { 
              assert_true(navigator.vibrate(1000), "vibrate() returns true when vibration is not supported");
          }, "Calling vibrate returns true");
      }
    </script>
  </body>
</html>
